Rename result.py -> datamanager.py
